requirements.txt
streamlit==1.32.0
pandas==2.1.4
gspread==5.12.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
plotly==5.20.0
numpy==1.26.4